fix ColumnIndexer to use long offsets patch by Thomas Richter; reviewed by jbellis for CASSANDRA-3358
git-svn-id: https://svn.apache.org/repos/asf/cassandra/branches/cassandra-0.7@1183000 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/cassandra/db/ColumnIndexer.java
@@ -70,7 +70,7 @@ public class ColumnIndexer
         // update bloom filter and create a list of IndexInfo objects marking the first and last column
         // in each block of ColumnIndexSize
         List<IndexHelper.IndexInfo> indexList = new ArrayList<IndexHelper.IndexInfo>();
-        int endPosition = 0, startPosition = -1;
+        long endPosition = 0, startPosition = -1;
         int indexSizeInBytes = 0;
         IColumn lastColumn = null, firstColumn = null;
         for (IColumn column : columns)
